add LICENSE file to project
The Functional Software copyright notice and text of the three clause
BSD license that this software is licensed under has been added to the
project root with a `LICENSE` file.

The `LICENSE` file contains the text of the license as hosted at
[docs.sentry.io/license](docs.sentry.io/license), with the copyright
date bumped to 2018 and the text formatting adjusted for an 80 column
plaintext layout.