fix(snuba): Add consumer group to transactions consumer
Without the consumer-group option, transactions consumer and events consumer will compete for messages on the same topic and usually events win, which is the first one. This may cause some data loss for performance and make it seem not work.

Should address https://forum.sentry.io/t/perfomance-tracing-for-sentry-itself/10405/5?u=byk