Work around slow Kafka startup
On some systems, Kafka takes a while to start up, which will error out
the relay service. For whichever reason, this isn't handled by the
dependencies expecting Kafka to be up before relay, which results in
relay service failing and never restarting.

This sets the relay service to always restart until it finds Kafka.