Resolve Safari Content-Disposition header bug
With release 23.7.0, the upstream sentry@23.7.0 upgraded Django to
3.2.20, a breaking change for that dependency. This introduced a number
of bugs into self-hosted, including one where the Content-Disposition
response header is now set by Django in a manner that is illegible to
Safari, as documenting in https://code.djangoproject.com/ticket/31082.

This change modifies the default nginx config to override this header
with the simpler `inline` setting, allowing Safari to properly
decompress gzipped resources like CSS and Javascript bundles.

Fixes #2285