fix(sentry-admin): Do not wait for command finish to display output
Currently sentry-admin.sh script saves the output of the command to a separate variable.
This makes the command "freeze" if it requires any input from the user (like sentry-admin.sh restore).
The change should provide the output directly to the shell.

Also changed contributing.md as it seemed outdated and updated requirements-dev.txt with the missing `cryptography` package.